+++ cli.go
@@ -18,4 +18,4 @@
 //   }
 package cli
 
-//go:generate python ./generate-flag-types -i flag-types.json -o flag_generated.go
+//go:generate python ./generate-flag-types cli -i flag-types.json -o flag_generated.go